uf2: Lock to 1st family ID found & add new ID range for IF updates.
As the Interface can accept more than one Family ID for target
UF2 blocks, it will only flash blocks for the first valid value
it finds.

A new UF2 range has been added for DAPLink interface images, so
the DAPLink bootloader can use those IDs for updating the
interface without accidentally flashing data intended for
the target (which can be an issue for boards like the micro:bit
V2.2x where the Interface and Target are the same MCU family).